+++ Dtos/CategoryDto.cs
@@ -5,11 +5,11 @@ namespace Dtos
 {
     public class CategoryDto
     {
-        public long Id { get; set; }
+        public int Id { get; set; }
         public string Name { get; set; }
         public string Description { get; set; }
 
-        public long? ParentId { get; set; }
+        public int? ParentId { get; set; }
         public bool HasParent { get; set; }
         public bool HasChild { get; set; }
         public List<CategoryFieldValueDto> CategoryFieldValues { get; set; }        
@@ -23,13 +23,13 @@ namespace Dtos
     }
     public class CategoryUpdateDto
     {
-        public long Id { get; set; }
+        public int Id { get; set; }
         public string Name { get; set; }
         public string Description { get; set; }
     }
 
     public class CategoryDeleteDto
     {
-        public long Id { get; set; }
+        public int Id { get; set; }
     }
 }
