+++ Dtos/CategoryFieldValueDto.cs
@@ -2,16 +2,16 @@ namespace Dtos
 {
     public class CategoryFieldValueDto
     {
-        public long FieldId { get; set; }
-        public long  CategoryId { get; set; }
+        public int FieldId { get; set; }
+        public int  CategoryId { get; set; }
         public string FieldName { get; set; }
         public string Value { get; set; }
     }
 
     public class CategoryFieldValueCreateDto
     {
-        public long FieldId { get; set; }
-        public long CategoryId { get; set; }
+        public int FieldId { get; set; }
+        public int CategoryId { get; set; }
         public string Value { get; set; }
     }
 }
